Fix controlers oauth2 sso setup
Ony one controllers needs to be annotated by @EnableOAuth2Sso so just provide
one for configuration and others will use that settings
