--- src/main/java/pik/controllers/GreetingController.java
@@ -1,26 +1,0 @@
-package pik.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@Controller
-public class GreetingController {
-
-
-    @RequestMapping("/greeting")
-    public String greeting(@RequestParam(value="name", required=false, defaultValue="World") String name, Model model) {
-        model.addAttribute("name", name);
-
-        return "greeting-test";
-    }
-
-    @RequestMapping("/testowy")
-    public String testowy(@RequestParam(value="name", required=false, defaultValue="World") String name, Model model) {
-        model.addAttribute("name", name);
-        return "greeting-test";
-    }
-
-
-}
